Move reflection-related utils from AssertUtil to new ReflectUtil.
This allows `AssertUtil` to link again on Scala.js.

Previously, the field of type `java.lang.Field` would pollute all
of `AssertUtil`, preventing any test using any of the methods of
`AssertUtil` from linking on Scala.js.
